usbphy_ll: Save delay measurement code
This approach will not be taken, but is interesting for
historical purposes.

At one point, the timer loop was to be self-calibrating.  That is, it
would measure the number of "ticks" between a high and a low pulse, and
adjust itself to match.

The problem is that measuring SysTick takes four cycles, meaning the
measured value can be off by 1-4 cycles.

So now we take the approach where the user tells us how many cycles.

Frequently, it's 32, and can be derived by the formula:

    F_CORE / 1500000

That is, the frequency of the CPU core (i.e. SysTick) divided by 1.5 MHz.

Signed-off-by: Sean Cross <xobs@kosagi.com>