usbphy: fix race condition in data transfer
Previously, we would set phy->data_is_queued before data was actually
queued.  If an interurpt hit, this could result in us sending out stale
data, because this bit is checked during the ISR to see if data is ready.

Set this bit after we queue data, to prevent this from happening.

Signed-off-by: Sean Cross <xobs@kosagi.com>