usbphy: remove spsave, bufsave from phy, change usbphyreadi
The usbPhyReadI() function now operates without clobbering the stack pointer,
so we no longer need to stash this in the usbphy struct.  As a bonus, this
means we can now make USBPHY const.

The function also no longer requires the buffer to be byte-aligned, so
change the signature to accept uint8_t* rather than uint32_t*.

Signed-off-by: Sean Cross <xobs@kosagi.com>